fix failing test related to withTimeoutOrNull (Job has not completed yet)
